♻️ (agent-core): share config rendering across View modal and wizard Review
Extract the per-architecture detail rendering into a shared AgentConfigView component, consumed by both the AgentCore View modal and all four wizard Review steps. The Review step now shows the same structured, human-readable view as Agent View instead of a raw JSON dump, with the configuration JSON available in a collapsed, copyable CodeView block.

Agents-as-tools references resolve to agent names (not runtime ARNs) via the runtimes list, matching the View modal. Slims view-version-modal by ~1000 lines by delegating its body to the shared renderer.
