+++ src/user-interface/react-app/src/components/admin/agent-core/view-version-modal.tsx
@@ -8,7 +8,6 @@ import {
     Box,
     BreadcrumbGroup,
     Button,
-    ColumnLayout,
     ExpandableSection,
     FormField,
     Link,
@@ -16,44 +15,14 @@ import {
     Select,
     SpaceBetween,
     StatusIndicator,
-    Table,
 } from "@cloudscape-design/components";
+import CodeView from "@cloudscape-design/code-view/code-view";
+import jsonHighlight from "@cloudscape-design/code-view/highlight/json";
 import CopyToClipboard from "@cloudscape-design/components/copy-to-clipboard";
-import { generateClient } from "aws-amplify/api";
-import { useContext, useEffect, useState } from "react";
-import { McpServer, RuntimeSummary } from "../../../API";
-import { AppContext } from "../../../common/app-context";
-import { listAvailableMcpServers as listAvailableMcpServersQuery } from "../../../graphql/queries";
-import GraphMinimap from "../../wizard/architectures/graph-minimap";
-import {
-    AgentCoreRuntimeConfiguration,
-    AgentsAsToolsConfiguration,
-    GraphConfiguration,
-    SwarmConfiguration,
-} from "../../wizard/types";
-
-const apiClient = generateClient();
-
-const isSwarmConfig = (config: any): config is SwarmConfiguration => {
-    return (
-        config &&
-        (Array.isArray(config.agents) || Array.isArray(config.agentReferences)) &&
-        typeof config.entryAgent === "string"
-    );
-};
-
-const isGraphConfig = (config: any): config is GraphConfiguration => {
-    return config && Array.isArray(config.nodes) && typeof config.entryPoint === "string";
-};
-
-const isAgentsAsToolsConfig = (config: any): config is AgentsAsToolsConfiguration => {
-    return (
-        config &&
-        Array.isArray(config.agentsAsTools) &&
-        config.modelInferenceParameters &&
-        typeof config.instructions === "string"
-    );
-};
+import { useEffect, useState } from "react";
+import { RuntimeSummary } from "../../../API";
+import { AgentCoreRuntimeConfiguration } from "../../wizard/types";
+import AgentConfigView, { AgentReferenceTarget } from "./agent-config-view";
 
 interface VersionInfo {
     version: string;
@@ -94,11 +63,8 @@ export default function ViewVersionModal({
     agents,
     onVersionSelect,
 }: ViewVersionModalProps) {
-    const appContext = useContext(AppContext);
-
     const [stack, setStack] = useState<NavFrame[]>([]);
     const [loadingConfig, setLoadingConfig] = useState(false);
-    const [availableMcpServers, setAvailableMcpServers] = useState<McpServer[]>([]);
 
     const current = stack.length > 0 ? stack[stack.length - 1] : null;
     const agentConfig = current?.config ?? null;
@@ -195,15 +161,21 @@ export default function ViewVersionModal({
     const goBack = () => setStack((prev) => prev.slice(0, -1));
     const popTo = (index: number) => setStack((prev) => prev.slice(0, index + 1));
 
-    // A reference is drillable only if it resolves to a runtime in this account.
-    // Otherwise (deleted, cross-account, raw runtimeId) render a non-clickable hint.
-    const renderAgentRef = (
-        displayName: string,
-        runtime: RuntimeSummary | null,
-        endpointName?: string,
-    ): React.ReactNode => {
+    // Turn an AgentConfigView reference into a drill-down link, or a non-clickable
+    // hint when the reference does not resolve to a runtime in this account (deleted,
+    // cross-account, or a raw runtimeId).
+    const renderAgentReference = (target: AgentReferenceTarget): React.ReactNode => {
+        const runtime = target.agentName
+            ? findRuntimeByName(target.agentName)
+            : target.runtimeId
+              ? findRuntimeById(target.runtimeId)
+              : null;
+        const displayName = target.runtimeId
+            ? getAgentNameByRuntimeId(target.runtimeId)
+            : target.display;
+
         if (runtime) {
-            return <Link onFollow={() => drillInto(runtime, endpointName)}>{displayName}</Link>;
+            return <Link onFollow={() => drillInto(runtime, target.endpoint)}>{displayName}</Link>;
         }
         return (
             <SpaceBetween direction="horizontal" size="xs" alignItems="center">
@@ -214,78 +186,6 @@ export default function ViewVersionModal({
             </SpaceBetween>
         );
     };
-
-    const getModelName = (modelId: string) => {
-        if (!appContext?.aws_bedrock_supported_models) return modelId;
-
-        const regionPrefix = appContext.aws_project_region.split("-")[0];
-
-        for (const [label, templateValue] of Object.entries(
-            appContext.aws_bedrock_supported_models,
-        )) {
-            const processedValue = templateValue.replace("[REGION-PREFIX]", regionPrefix);
-            if (processedValue === modelId) {
-                return label;
-            }
-        }
-
-        return modelId; // fallback to showing the ID if no match found
-    };
-
-    const isThinkingEnabled = (reasoningBudget: number | string | null | undefined): boolean => {
-        return (
-            reasoningBudget !== undefined &&
-            reasoningBudget !== null &&
-            reasoningBudget !== "disabled" &&
-            reasoningBudget !== 0
-        );
-    };
-
-    const renderThinkingStatus = (reasoningBudget: number | string | null | undefined) => {
-        if (isThinkingEnabled(reasoningBudget)) {
-            const budgetLabel =
-                typeof reasoningBudget === "number"
-                    ? `${reasoningBudget} tokens`
-                    : String(reasoningBudget);
-            return (
-                <SpaceBetween direction="vertical" size="xxs">
-                    <StatusIndicator type="success">Enabled</StatusIndicator>
-                    <Box variant="awsui-key-label" display="inline">
-                        Budget:{" "}
-                        <Box display="inline" fontWeight="normal">
-                            {budgetLabel}
-                        </Box>
-                    </Box>
-                </SpaceBetween>
-            );
-        }
-        return <StatusIndicator type="stopped">Disabled</StatusIndicator>;
-    };
-
-    const renderMemoryStatus = (useMemory: boolean | undefined) => {
-        if (useMemory) {
-            return <StatusIndicator type="success">Attached</StatusIndicator>;
-        }
-        return <StatusIndicator type="stopped">Not attached</StatusIndicator>;
-    };
-
-    // Fetch available MCP servers when modal opens
-    useEffect(() => {
-        const fetchMcpServers = async () => {
-            try {
-                const result = await apiClient.graphql({ query: listAvailableMcpServersQuery });
-                if (result.data?.listAvailableMcpServers) {
-                    setAvailableMcpServers(result.data.listAvailableMcpServers as McpServer[]);
-                }
-            } catch (error) {
-                console.error("Failed to fetch MCP servers:", error);
-            }
-        };
-
-        if (visible) {
-            fetchMcpServers();
-        }
-    }, [visible]);
 
     // Initialize the navigation stack with the opened (root) agent.
     useEffect(() => {
@@ -320,128 +220,7 @@ export default function ViewVersionModal({
         onDismiss();
     };
 
-    // Recursive function to render parameter values
-    const renderValue = (value: unknown, depth: number = 0): React.ReactNode => {
-        if (value === null || value === undefined) {
-            return <Box color="text-status-inactive">null</Box>;
-        }
-
-        if (Array.isArray(value)) {
-            if (value.length === 0) {
-                return <Box color="text-status-inactive">[]</Box>;
-            }
-            return (
-                <SpaceBetween direction="vertical" size="xxs">
-                    {value.map((item, index) => (
-                        <Box key={index} margin={{ left: depth > 0 ? "l" : undefined }}>
-                            <Box variant="awsui-key-label" display="inline">
-                                [{index}]:
-                            </Box>{" "}
-                            {renderValue(item, depth + 1)}
-                        </Box>
-                    ))}
-                </SpaceBetween>
-            );
-        }
-
-        if (typeof value === "object") {
-            const entries = Object.entries(value as Record<string, unknown>);
-            if (entries.length === 0) {
-                return <Box color="text-status-inactive">{"{}"}</Box>;
-            }
-            return (
-                <div style={{ display: "flex", flexDirection: "column", gap: "2px" }}>
-                    {entries.map(([key, val]) => (
-                        <div key={key} style={{ marginLeft: depth > 0 ? "16px" : undefined }}>
-                            {typeof val === "object" && val !== null ? (
-                                <>
-                                    <Box variant="awsui-key-label">{key}:</Box>
-                                    <div style={{ marginTop: "2px" }}>
-                                        {renderValue(val, depth + 1)}
-                                    </div>
-                                </>
-                            ) : (
-                                <span>
-                                    <Box variant="awsui-key-label" display="inline">
-                                        {key}:
-                                    </Box>{" "}
-                                    {renderValue(val, depth + 1)}
-                                </span>
-                            )}
-                        </div>
-                    ))}
-                </div>
-            );
-        }
-
-        if (typeof value === "boolean") {
-            return (
-                <Box display="inline" color="text-status-info">
-                    {value.toString()}
-                </Box>
-            );
-        }
-
-        if (typeof value === "number") {
-            return <Box display="inline">{value}</Box>;
-        }
-
-        return <Box display="inline">{String(value)}</Box>;
-    };
-
-    // Map a list of MCP server names to display rows with descriptions.
-    const buildMcpItems = (servers: string[]) =>
-        servers.map((serverName) => {
-            const serverInfo = availableMcpServers.find((s) => s.name === serverName);
-            return {
-                name: serverName,
-                description: serverInfo?.description || "No description available",
-            };
-        });
-
-    type McpServerRow = { name: string; description: string };
-    const mcpServerColumns = [
-        {
-            id: "name",
-            header: "Server Name",
-            cell: (item: McpServerRow) => item.name,
-            isRowHeader: true,
-        },
-        {
-            id: "description",
-            header: "Description",
-            cell: (item: McpServerRow) => (
-                <Box
-                    color={
-                        item.description === "No description available"
-                            ? "text-status-inactive"
-                            : undefined
-                    }
-                >
-                    {item.description}
-                </Box>
-            ),
-        },
-    ];
-
-    const isSwarm = agentConfig && isSwarmConfig(agentConfig);
-    const isAgentsAsTools =
-        agentConfig &&
-        !isSwarm &&
-        !isGraphConfig(agentConfig) &&
-        isAgentsAsToolsConfig(agentConfig);
-
-    const toolTableItems =
-        !isSwarm && agentConfig?.tools
-            ? agentConfig.tools.map((toolName) => ({
-                  name: toolName,
-                  parameters: agentConfig.toolParameters[toolName] || {},
-              }))
-            : [];
-
-    // Get MCP server details for configured servers (single-agent branch)
-    const mcpServerTableItems =
-        !isSwarm && agentConfig?.mcpServers ? buildMcpItems(agentConfig.mcpServers) : [];
+    const rawJson = agentConfig ? JSON.stringify(agentConfig, null, 2) : "";
 
     return (
         <Modal
@@ -542,788 +321,31 @@ export default function ViewVersionModal({
                         <StatusIndicator type="loading">Loading configuration</StatusIndicator>
                     </Box>
                 ) : agentConfig ? (
-                    isSwarmConfig(agentConfig) ? (
-                        <SpaceBetween direction="vertical" size="m">
-                            <FormField label="Entry Agent">
-                                <Box padding="m">{agentConfig.entryAgent}</Box>
-                            </FormField>
+                    <>
+                        <AgentConfigView
+                            config={agentConfig}
+                            renderAgentReference={renderAgentReference}
+                        />
 
-                            <FormField label="AgentCore Memory">
-                                <Box padding="m">{renderMemoryStatus(agentConfig.useMemory)}</Box>
-                            </FormField>
-
-                            {agentConfig.agents && agentConfig.agents.length > 0 && (
-                                <FormField label="Inline Agents">
-                                    <SpaceBetween direction="vertical" size="xs">
-                                        {agentConfig.agents.map((inlineAgent, idx) => {
-                                            const params =
-                                                inlineAgent.modelInferenceParameters?.parameters;
-                                            const inlineMcp = buildMcpItems(
-                                                inlineAgent.mcpServers || [],
-                                            );
-                                            const inlineToolParams =
-                                                inlineAgent.toolParameters || {};
-                                            return (
-                                                <ExpandableSection
-                                                    key={idx}
-                                                    variant="container"
-                                                    headerText={inlineAgent.name}
-                                                >
-                                                    <SpaceBetween direction="vertical" size="m">
-                                                        <ColumnLayout
-                                                            columns={4}
-                                                            variant="text-grid"
-                                                        >
-                                                            <div>
-                                                                <Box variant="awsui-key-label">
-                                                                    Model
-                                                                </Box>
-                                                                <Box>
-                                                                    {getModelName(
-                                                                        inlineAgent
-                                                                            .modelInferenceParameters
-                                                                            ?.modelId || "",
-                                                                    )}
-                                                                </Box>
-                                                            </div>
-                                                            <div>
-                                                                <Box variant="awsui-key-label">
-                                                                    Temperature
-                                                                </Box>
-                                                                <Box>
-                                                                    {params?.temperature ?? "N/A"}
-                                                                </Box>
-                                                            </div>
-                                                            <div>
-                                                                <Box variant="awsui-key-label">
-                                                                    Max Tokens
-                                                                </Box>
-                                                                <Box>
-                                                                    {params?.maxTokens ?? "N/A"}
-                                                                </Box>
-                                                            </div>
-                                                            <div>
-                                                                <Box variant="awsui-key-label">
-                                                                    Thinking
-                                                                </Box>
-                                                                <Box>
-                                                                    {renderThinkingStatus(
-                                                                        inlineAgent
-                                                                            .modelInferenceParameters
-                                                                            ?.reasoningBudget,
-                                                                    )}
-                                                                </Box>
-                                                            </div>
-                                                        </ColumnLayout>
-
-                                                        <div>
-                                                            <Box variant="awsui-key-label">
-                                                                Tools
-                                                            </Box>
-                                                            <Box>
-                                                                {inlineAgent.tools?.length > 0
-                                                                    ? inlineAgent.tools.join(", ")
-                                                                    : "None"}
-                                                            </Box>
-                                                        </div>
-
-                                                        {inlineMcp.length > 0 && (
-                                                            <div>
-                                                                <Box variant="awsui-key-label">
-                                                                    MCP Servers
-                                                                </Box>
-                                                                <Table
-                                                                    variant="embedded"
-                                                                    items={inlineMcp}
-                                                                    columnDefinitions={
-                                                                        mcpServerColumns
-                                                                    }
-                                                                />
-                                                            </div>
-                                                        )}
-
-                                                        {Object.keys(inlineToolParams).length >
-                                                            0 && (
-                                                            <div>
-                                                                <Box variant="awsui-key-label">
-                                                                    Tool Parameters
-                                                                </Box>
-                                                                {renderValue(inlineToolParams)}
-                                                            </div>
-                                                        )}
-
-                                                        <FormField
-                                                            label={
-                                                                <SpaceBetween
-                                                                    direction="horizontal"
-                                                                    size="xs"
-                                                                    alignItems="center"
-                                                                >
-                                                                    <span>Instructions</span>
-                                                                    <CopyToClipboard
-                                                                        textToCopy={
-                                                                            inlineAgent.instructions ||
-                                                                            ""
-                                                                        }
-                                                                        variant="icon"
-                                                                        copySuccessText="Instructions copied"
-                                                                        copyErrorText="Failed to copy"
-                                                                    />
-                                                                </SpaceBetween>
-                                                            }
-                                                        >
-                                                            <ExpandableSection
-                                                                headerText="Show instructions"
-                                                                defaultExpanded={false}
-                                                            >
-                                                                <Box padding="m" variant="code">
-                                                                    <pre
-                                                                        style={{
-                                                                            margin: 0,
-                                                                            whiteSpace: "pre-wrap",
-                                                                            wordWrap: "break-word",
-                                                                        }}
-                                                                    >
-                                                                        {inlineAgent.instructions}
-                                                                    </pre>
-                                                                </Box>
-                                                            </ExpandableSection>
-                                                        </FormField>
-                                                    </SpaceBetween>
-                                                </ExpandableSection>
-                                            );
-                                        })}
-                                    </SpaceBetween>
-                                </FormField>
-                            )}
-
-                            {agentConfig.agentReferences &&
-                                agentConfig.agentReferences.length > 0 && (
-                                    <FormField label="Agent References">
-                                        <Table
-                                            items={agentConfig.agentReferences}
-                                            columnDefinitions={[
-                                                {
-                                                    id: "agentName",
-                                                    header: "Agent Name",
-                                                    cell: (item: any) =>
-                                                        renderAgentRef(
-                                                            item.agentName,
-                                                            findRuntimeByName(item.agentName),
-                                                            item.endpointName,
-                                                        ),
-                                                    isRowHeader: true,
-                                                },
-                                                {
-                                                    id: "endpointName",
-                                                    header: "Endpoint",
-                                                    cell: (item: any) => item.endpointName,
-                                                },
-                                            ]}
-                                        />
-                                    </FormField>
-                                )}
-
-                            {agentConfig.orchestrator && (
-                                <FormField label="Orchestrator Settings">
-                                    <Box padding="m">
-                                        <ColumnLayout columns={4} variant="text-grid">
-                                            <div>
-                                                <Box variant="awsui-key-label">Max Handoffs</Box>
-                                                <Box>
-                                                    {agentConfig.orchestrator.maxHandoffs ?? "N/A"}
-                                                </Box>
-                                            </div>
-                                            <div>
-                                                <Box variant="awsui-key-label">Max Iterations</Box>
-                                                <Box>
-                                                    {agentConfig.orchestrator.maxIterations ??
-                                                        "N/A"}
-                                                </Box>
-                                            </div>
-                                            <div>
-                                                <Box variant="awsui-key-label">
-                                                    Execution Timeout
-                                                </Box>
-                                                <Box>
-                                                    {agentConfig.orchestrator
-                                                        .executionTimeoutSeconds ?? "N/A"}
-                                                    s
-                                                </Box>
-                                            </div>
-                                            <div>
-                                                <Box variant="awsui-key-label">Node Timeout</Box>
-                                                <Box>
-                                                    {agentConfig.orchestrator.nodeTimeoutSeconds ??
-                                                        "N/A"}
-                                                    s
-                                                </Box>
-                                            </div>
-                                        </ColumnLayout>
-                                    </Box>
-                                </FormField>
-                            )}
-
-                            {agentConfig.conversationManager && (
-                                <FormField label="Conversation Manager">
-                                    <Box padding="m">{agentConfig.conversationManager}</Box>
-                                </FormField>
-                            )}
-                        </SpaceBetween>
-                    ) : isAgentsAsTools ? (
-                        <SpaceBetween direction="vertical" size="m">
-                            <FormField label="Model Configuration">
-                                <Box padding="m">
-                                    <ColumnLayout columns={4} variant="text-grid">
-                                        <div>
-                                            <Box variant="awsui-key-label">Model</Box>
-                                            <Box>
-                                                {getModelName(
-                                                    agentConfig.modelInferenceParameters?.modelId ||
-                                                        "",
-                                                )}
-                                            </Box>
-                                        </div>
-                                        <div>
-                                            <Box variant="awsui-key-label">Temperature</Box>
-                                            <Box>
-                                                {agentConfig.modelInferenceParameters?.parameters
-                                                    ?.temperature ?? "N/A"}
-                                            </Box>
-                                        </div>
-                                        <div>
-                                            <Box variant="awsui-key-label">Max Tokens</Box>
-                                            <Box>
-                                                {agentConfig.modelInferenceParameters?.parameters
-                                                    ?.maxTokens ?? "N/A"}
-                                            </Box>
-                                        </div>
-                                        <div>
-                                            <Box variant="awsui-key-label">Thinking</Box>
-                                            <Box>
-                                                {renderThinkingStatus(
-                                                    agentConfig.modelInferenceParameters
-                                                        ?.reasoningBudget,
-                                                )}
-                                            </Box>
-                                        </div>
-                                    </ColumnLayout>
-                                </Box>
-                            </FormField>
-
-                            <FormField label="AgentCore Memory">
-                                <Box padding="m">{renderMemoryStatus(agentConfig.useMemory)}</Box>
-                            </FormField>
-
-                            <FormField
-                                label={
-                                    <SpaceBetween
-                                        direction="horizontal"
-                                        size="xs"
-                                        alignItems="center"
-                                    >
-                                        <span>Orchestrator Instructions</span>
-                                        <CopyToClipboard
-                                            textToCopy={agentConfig.instructions || ""}
-                                            variant="icon"
-                                            copySuccessText="Instructions copied"
-                                            copyErrorText="Failed to copy"
-                                        />
-                                    </SpaceBetween>
+                        <ExpandableSection
+                            headerText="Configuration JSON"
+                            defaultExpanded={false}
+                        >
+                            <CodeView
+                                content={rawJson}
+                                highlight={jsonHighlight}
+                                wrapLines
+                                actions={
+                                    <CopyToClipboard
+                                        variant="icon"
+                                        textToCopy={rawJson}
+                                        copySuccessText="Configuration copied"
+                                        copyErrorText="Failed to copy"
+                                    />
                                 }
-                            >
-                                <ExpandableSection
-                                    headerText="Show instructions"
-                                    defaultExpanded={false}
-                                >
-                                    <Box padding="m" variant="code">
-                                        <pre
-                                            style={{
-                                                margin: 0,
-                                                whiteSpace: "pre-wrap",
-                                                wordWrap: "break-word",
-                                            }}
-                                        >
-                                            {agentConfig.instructions}
-                                        </pre>
-                                    </Box>
-                                </ExpandableSection>
-                            </FormField>
-
-                            {agentConfig.agentsAsTools && agentConfig.agentsAsTools.length > 0 && (
-                                <FormField label="Agents as Tools">
-                                    <Table
-                                        items={agentConfig.agentsAsTools}
-                                        columnDefinitions={[
-                                            {
-                                                id: "agentName",
-                                                header: "Agent",
-                                                cell: (item: any) =>
-                                                    renderAgentRef(
-                                                        getAgentNameByRuntimeId(item.runtimeId),
-                                                        findRuntimeById(item.runtimeId),
-                                                        item.endpoint,
-                                                    ),
-                                                isRowHeader: true,
-                                            },
-                                            {
-                                                id: "endpoint",
-                                                header: "Endpoint",
-                                                cell: (item: any) => item.endpoint || "DEFAULT",
-                                            },
-                                        ]}
-                                    />
-                                </FormField>
-                            )}
-
-                            {agentConfig.tools && agentConfig.tools.length > 0 && (
-                                <FormField label="Additional Tools">
-                                    <Table
-                                        items={agentConfig.tools.map((toolName: string) => ({
-                                            name: toolName,
-                                            parameters:
-                                                agentConfig.toolParameters?.[toolName] || {},
-                                        }))}
-                                        columnDefinitions={[
-                                            {
-                                                id: "name",
-                                                header: "Tool Name",
-                                                cell: (item: any) => item.name,
-                                                isRowHeader: true,
-                                            },
-                                            {
-                                                id: "parameters",
-                                                header: "Parameters",
-                                                cell: (item: any) => {
-                                                    if (Object.keys(item.parameters).length === 0) {
-                                                        return (
-                                                            <Box color="text-status-inactive">
-                                                                None
-                                                            </Box>
-                                                        );
-                                                    }
-                                                    return renderValue(item.parameters);
-                                                },
-                                            },
-                                        ]}
-                                    />
-                                </FormField>
-                            )}
-
-                            {agentConfig.mcpServers && agentConfig.mcpServers.length > 0 && (
-                                <FormField label="MCP Servers">
-                                    <Table
-                                        items={buildMcpItems(agentConfig.mcpServers)}
-                                        columnDefinitions={mcpServerColumns}
-                                    />
-                                </FormField>
-                            )}
-
-                            {agentConfig.conversationManager && (
-                                <FormField label="Conversation Manager">
-                                    <Box padding="m">{agentConfig.conversationManager}</Box>
-                                </FormField>
-                            )}
-                        </SpaceBetween>
-                    ) : isGraphConfig(agentConfig) ? (
-                        <SpaceBetween direction="vertical" size="m">
-                            <FormField label="Entry Point">
-                                <Box padding="m">{agentConfig.entryPoint}</Box>
-                            </FormField>
-
-                            <FormField label="AgentCore Memory">
-                                <Box padding="m">{renderMemoryStatus(agentConfig.useMemory)}</Box>
-                            </FormField>
-
-                            {agentConfig.nodes && agentConfig.nodes.length > 0 && (
-                                <FormField label="Graph Nodes">
-                                    <Table
-                                        items={agentConfig.nodes}
-                                        columnDefinitions={[
-                                            {
-                                                id: "id",
-                                                header: "Node ID",
-                                                cell: (item: any) => item.id,
-                                                isRowHeader: true,
-                                            },
-                                            {
-                                                id: "agentName",
-                                                header: "Agent",
-                                                cell: (item: any) =>
-                                                    item.agentName ? (
-                                                        renderAgentRef(
-                                                            item.agentName,
-                                                            findRuntimeByName(item.agentName),
-                                                            item.endpointName,
-                                                        )
-                                                    ) : (
-                                                        <Box color="text-status-inactive">
-                                                            {item.deterministicNodeKey ||
-                                                                item.nodeType ||
-                                                                "-"}
-                                                        </Box>
-                                                    ),
-                                            },
-                                            {
-                                                id: "endpointName",
-                                                header: "Endpoint",
-                                                cell: (item: any) => item.endpointName || "DEFAULT",
-                                            },
-                                            {
-                                                id: "label",
-                                                header: "Label",
-                                                cell: (item: any) => item.label || "-",
-                                            },
-                                            {
-                                                id: "dynamicMap",
-                                                header: "Dynamic Map",
-                                                cell: (item: any) =>
-                                                    item.dynamicMapConfig ? (
-                                                        <Box fontSize="body-s">
-                                                            {item.dynamicMapConfig.sourceKey} →{" "}
-                                                            {item.dynamicMapConfig.targetNode}
-                                                        </Box>
-                                                    ) : (
-                                                        <Box color="text-status-inactive">-</Box>
-                                                    ),
-                                            },
-                                            {
-                                                id: "promptTemplate",
-                                                header: "Prompt Template",
-                                                cell: (item: any) =>
-                                                    item.promptTemplate ? (
-                                                        <span
-                                                            title={item.promptTemplate}
-                                                            style={{
-                                                                display: "-webkit-box",
-                                                                WebkitLineClamp: 2,
-                                                                WebkitBoxOrient:
-                                                                    "vertical" as const,
-                                                                overflow: "hidden",
-                                                                textOverflow: "ellipsis",
-                                                                whiteSpace: "pre-wrap",
-                                                                maxWidth: 250,
-                                                                fontSize: 12,
-                                                            }}
-                                                        >
-                                                            {item.promptTemplate}
-                                                        </span>
-                                                    ) : (
-                                                        <Box color="text-status-inactive">-</Box>
-                                                    ),
-                                            },
-                                        ]}
-                                    />
-                                </FormField>
-                            )}
-
-                            {agentConfig.edges && agentConfig.edges.length > 0 && (
-                                <FormField label="Edges">
-                                    <Table
-                                        items={agentConfig.edges}
-                                        columnDefinitions={[
-                                            {
-                                                id: "source",
-                                                header: "Source",
-                                                cell: (item: any) => item.source,
-                                                isRowHeader: true,
-                                            },
-                                            {
-                                                id: "target",
-                                                header: "Target",
-                                                cell: (item: any) => item.target,
-                                            },
-                                            {
-                                                id: "condition",
-                                                header: "Condition",
-                                                cell: (item: any) =>
-                                                    item.condition || (
-                                                        <Box color="text-status-inactive">-</Box>
-                                                    ),
-                                            },
-                                        ]}
-                                    />
-                                </FormField>
-                            )}
-
-                            <FormField label="Graph State">
-                                <Box padding="m">
-                                    {agentConfig.stateClass ? (
-                                        <div>
-                                            <Box variant="awsui-key-label">State Class</Box>
-                                            <Box>{agentConfig.stateClass}</Box>
-                                        </div>
-                                    ) : agentConfig.stateSchema &&
-                                      Object.keys(agentConfig.stateSchema).length > 0 ? (
-                                        <SpaceBetween direction="vertical" size="xs">
-                                            {Object.entries(agentConfig.stateSchema).map(
-                                                ([key, type]) => (
-                                                    <Box key={key}>
-                                                        <Box
-                                                            variant="awsui-key-label"
-                                                            display="inline"
-                                                        >
-                                                            {key}:
-                                                        </Box>{" "}
-                                                        <Box display="inline">{String(type)}</Box>
-                                                    </Box>
-                                                ),
-                                            )}
-                                        </SpaceBetween>
-                                    ) : (
-                                        <Box color="text-status-inactive">Not configured</Box>
-                                    )}
-                                </Box>
-                            </FormField>
-
-                            <FormField label="Graph Topology">
-                                <GraphMinimap
-                                    graphConfig={{
-                                        nodes: agentConfig.nodes || [],
-                                        edges: agentConfig.edges || [],
-                                        entryPoint: agentConfig.entryPoint,
-                                        stateSchema: agentConfig.stateSchema || {},
-                                        stateClass: agentConfig.stateClass,
-                                        orchestrator: agentConfig.orchestrator || {
-                                            maxIterations: 50,
-                                            executionTimeoutSeconds: 300,
-                                            nodeTimeoutSeconds: 60,
-                                        },
-                                    }}
-                                />
-                            </FormField>
-
-                            {agentConfig.orchestrator && (
-                                <FormField label="Orchestrator Settings">
-                                    <Box padding="m">
-                                        <ColumnLayout columns={3} variant="text-grid">
-                                            <div>
-                                                <Box variant="awsui-key-label">Max Iterations</Box>
-                                                <Box>
-                                                    {agentConfig.orchestrator.maxIterations ??
-                                                        "N/A"}
-                                                </Box>
-                                            </div>
-                                            <div>
-                                                <Box variant="awsui-key-label">
-                                                    Execution Timeout
-                                                </Box>
-                                                <Box>
-                                                    {agentConfig.orchestrator
-                                                        .executionTimeoutSeconds ?? "N/A"}
-                                                    s
-                                                </Box>
-                                            </div>
-                                            <div>
-                                                <Box variant="awsui-key-label">Node Timeout</Box>
-                                                <Box>
-                                                    {agentConfig.orchestrator.nodeTimeoutSeconds ??
-                                                        "N/A"}
-                                                    s
-                                                </Box>
-                                            </div>
-                                        </ColumnLayout>
-                                    </Box>
-                                </FormField>
-                            )}
-                        </SpaceBetween>
-                    ) : (
-                        <SpaceBetween direction="vertical" size="m">
-                            <FormField label="Model Configuration">
-                                <Box padding="m">
-                                    <ColumnLayout columns={4} variant="text-grid">
-                                        <div>
-                                            <Box variant="awsui-key-label">Model</Box>
-                                            <Box>
-                                                {getModelName(
-                                                    agentConfig.modelInferenceParameters.modelId,
-                                                )}
-                                            </Box>
-                                        </div>
-                                        <div>
-                                            <Box variant="awsui-key-label">Temperature</Box>
-                                            <Box>
-                                                {agentConfig.modelInferenceParameters.parameters
-                                                    .temperature ?? "N/A"}
-                                            </Box>
-                                        </div>
-                                        <div>
-                                            <Box variant="awsui-key-label">Max Tokens</Box>
-                                            <Box>
-                                                {agentConfig.modelInferenceParameters.parameters
-                                                    .maxTokens ?? "N/A"}
-                                            </Box>
-                                        </div>
-                                        <div>
-                                            <Box variant="awsui-key-label">Thinking</Box>
-                                            <Box>
-                                                {renderThinkingStatus(
-                                                    agentConfig.modelInferenceParameters
-                                                        .reasoningBudget,
-                                                )}
-                                            </Box>
-                                        </div>
-                                    </ColumnLayout>
-                                </Box>
-                            </FormField>
-
-                            <FormField label="AgentCore Memory">
-                                <Box padding="m">{renderMemoryStatus(agentConfig.useMemory)}</Box>
-                            </FormField>
-
-                            <FormField
-                                label="Agent Description"
-                                description="Capability blurb published in this agent's A2A agent card."
-                            >
-                                <Box padding="m">
-                                    {agentConfig.description || (
-                                        <Box color="text-status-inactive">Not set</Box>
-                                    )}
-                                </Box>
-                            </FormField>
-
-                            <FormField
-                                label={
-                                    <SpaceBetween
-                                        direction="horizontal"
-                                        size="xs"
-                                        alignItems="center"
-                                    >
-                                        <span>Agent Instructions</span>
-                                        <CopyToClipboard
-                                            textToCopy={agentConfig.instructions}
-                                            variant="icon"
-                                            copySuccessText="Instructions copied"
-                                            copyErrorText="Failed to copy"
-                                        />
-                                    </SpaceBetween>
-                                }
-                            >
-                                <ExpandableSection
-                                    headerText="Show instructions"
-                                    defaultExpanded={false}
-                                >
-                                    <Box padding="m" variant="code">
-                                        <pre
-                                            style={{
-                                                margin: 0,
-                                                whiteSpace: "pre-wrap",
-                                                wordWrap: "break-word",
-                                            }}
-                                        >
-                                            {agentConfig.instructions}
-                                        </pre>
-                                    </Box>
-                                </ExpandableSection>
-                            </FormField>
-
-                            {toolTableItems.length > 0 && (
-                                <FormField label="Tools and Parameters">
-                                    <Table
-                                        items={toolTableItems}
-                                        columnDefinitions={[
-                                            {
-                                                id: "name",
-                                                header: "Tool Name",
-                                                cell: (item) => item.name,
-                                                isRowHeader: true,
-                                            },
-                                            {
-                                                id: "parameters",
-                                                header: "Parameters",
-                                                cell: (item) => {
-                                                    if (Object.keys(item.parameters).length === 0) {
-                                                        return (
-                                                            <Box color="text-status-inactive">
-                                                                None
-                                                            </Box>
-                                                        );
-                                                    }
-                                                    return renderValue(item.parameters);
-                                                },
-                                            },
-                                        ]}
-                                    />
-                                </FormField>
-                            )}
-
-                            {agentConfig.skills && agentConfig.skills.length > 0 && (
-                                <FormField label="Skills">
-                                    <Box padding="m">{agentConfig.skills.join(", ")}</Box>
-                                </FormField>
-                            )}
-
-                            {mcpServerTableItems.length > 0 && (
-                                <FormField label="MCP Servers">
-                                    <Table
-                                        items={mcpServerTableItems}
-                                        columnDefinitions={mcpServerColumns}
-                                    />
-                                </FormField>
-                            )}
-
-                            <FormField label="Structured Output">
-                                {Array.isArray(agentConfig.structuredOutput) &&
-                                agentConfig.structuredOutput.length > 0 ? (
-                                    <Table
-                                        items={agentConfig.structuredOutput}
-                                        columnDefinitions={[
-                                            {
-                                                id: "name",
-                                                header: "Field Name",
-                                                cell: (item) => item.name,
-                                                isRowHeader: true,
-                                            },
-                                            {
-                                                id: "pythonType",
-                                                header: "Python Type",
-                                                cell: (item) => (
-                                                    <span style={{ fontFamily: "monospace" }}>
-                                                        {item.pythonType}
-                                                    </span>
-                                                ),
-                                            },
-                                            {
-                                                id: "description",
-                                                header: "Description",
-                                                cell: (item) => (
-                                                    <span style={{ whiteSpace: "pre-wrap" }}>
-                                                        {item.description}
-                                                    </span>
-                                                ),
-                                            },
-                                            {
-                                                id: "optional",
-                                                header: "Optional",
-                                                cell: (item) =>
-                                                    item.optional ? (
-                                                        <StatusIndicator type="info">
-                                                            Yes
-                                                        </StatusIndicator>
-                                                    ) : (
-                                                        <StatusIndicator type="stopped">
-                                                            No
-                                                        </StatusIndicator>
-                                                    ),
-                                            },
-                                        ]}
-                                    />
-                                ) : (
-                                    <Box padding="m">
-                                        <StatusIndicator type="stopped">
-                                            Not configured
-                                        </StatusIndicator>
-                                    </Box>
-                                )}
-                            </FormField>
-
-                            {agentConfig.conversationManager && (
-                                <FormField label="Conversation Manager">
-                                    <Box padding="m">{agentConfig.conversationManager}</Box>
-                                </FormField>
-                            )}
-                        </SpaceBetween>
-                    )
+                            />
+                        </ExpandableSection>
+                    </>
                 ) : null}
             </SpaceBetween>
         </Modal>
